% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_as_sf.R
\name{coords_as_sf}
\alias{coords_as_sf}
\title{Coerce dataframe with coordinates to an \code{sf} object}
\usage{
coords_as_sf(df, x, y, crs = "wgs84")
}
\arguments{
\item{df}{A data frame containing coordinate information}

\item{x}{A character string coresponding to column name in \code{df} containing the 'x' coordinate of the waypoint (aka. Longitude).}

\item{y}{A character string coresponding to column name in \code{df} containing the 'y' coordinate of the waypoint (aka. Latitude).}

\item{crs}{A character string coresponding to coordinate reference system of the waypoint. Defaults to "wgs84", use this for unprojected coordinates (Longitude & Latitude).}
}
\value{
A dataframe with \code{sf} geometry column appended
}
\description{
Convert dataframe with coordinate infromation into a plottable 'sf' style object
}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
