% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_markup.R
\name{delete_markup}
\alias{delete_markup}
\title{Delete Markup file from OnX Web Maps}
\usage{
delete_markup(uuid, token = Sys.getenv("ONX_TOKEN", unset = ""))
}
\arguments{
\item{uuid}{A character string containing the UUID of the markup file to be deleted. This can be found running \code{ronxmaps::onx_list_markups(parse = T)} and filtering by name.}

\item{token}{A Character string corresponding to the bearer token from the OnX OAuth/Ory session. Derived from \code{getToken}, typically stored in \code{Sys.getenv("ONX_TOKEN")}.}
}
\value{
Message if delete request was successful or HTTP error code if failed.
}
\description{
Use valid OnX Web Maps access token to delete a markup file from the web app via the API.
}
\note{
If the token is stale, then this function will not work. Refresh by re-running \code{getToken} then re-run function.
}
\examples{
## Not Run:
markup_lst <- ronxmaps::onx_list_markups(parse = T)
markup_toDelete <- markup_lst[which(markup_lst$name == "Borah Peak"),]
delete_markup(uuid = markup_toDelete$uuid, token = Sys.getenv("ONX_TOKEN", unset = ""))
## End Not Run

}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
