% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Extract OnX Web Maps Access Token}
\usage{
getToken(
  onx_email = Sys.getenv("ONX_EMAIL"),
  onx_password = Sys.getenv("ONX_PASSWORD"),
  url = onx_login_base(),
  verbose = T
)
}
\arguments{
\item{onx_email}{A string containing a valid OnX account email address. Best to store in environment for privacy (\verb{Sys.setenv("ONX_EMAIL" = "example@email.com"}).}

\item{onx_password}{A string containing a valid password to an active OnX account. Best to store in environment for privacy (\verb{Sys.setenv("ONX_PASSWORD" = "MyOnXpassWord123"}).}

\item{url}{A string containing a valid URL to the OnX Web Maps login page. Defaults to internal function \code{onx_login_base()}.}

\item{verbose}{A boolean indicating whether to return message stating how long the access token is valid (minutes). New tokens are typically valid for ~20 minutes. If token becomes stale, then this function will have to be re-run to get a fresh token.}
}
\value{
Access token is saved to environment session for future use in API function via \code{Sys.getenv("ONX_TOKEN")}. The system time of when the current token expires is stored in \code{Sys.getenv("ONX_TOKEN_EXP")}.
}
\description{
Login to OnX Web Maps, extract access token, and save token to environment session for use in API functions.
}
\note{
A fresh token will only be generated if the previous token is stale. If unsure, re-running this function \code{getToken(verbose=T)} will return how many minutes the current token will remain valid.
}
\examples{
## Not Run:
Sys.setenv(ONX_EMAIL = onx_email)
Sys.setenv(ONX_PASSWORD = onx_password)
getToken(verbose = T)
Sys.getenv("ONX_TOKEN")
Sys.getenv("ONX_TOKEN_EXP")
## End Not Run

}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
