% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kml2OnX.R
\name{import_kml2OnX}
\alias{import_kml2OnX}
\title{Import KML file to OnX Web Maps}
\usage{
import_kml2OnX(file, token = Sys.getenv("ONX_TOKEN", unset = ""))
}
\arguments{
\item{file}{A string containing the full file path to the KML file. Shapefiles in R must first be saved as KML files (ex., \code{sf::st_write(shp_file,"file.kml")}).}

\item{token}{Character bearer token from the OnX OAuth/Ory session. Derived from \code{getToken}, typically stored in \code{Sys.getenv("ONX_TOKEN")}.}
}
\value{
Markup ID of successfully uploaded file. If upload fails, the HTTP error code is returned in a message.
}
\description{
Use valid OnX Web Maps access token to import a KML file to the web app via the API.
}
\note{
If the token is stale, then this function will not work. Refresh by re-running \code{getToken} then re-run function.
}
\examples{
## Not Run:
test_file <- ronxmaps::BorahPeak_kml
sf::st_write(test_file,"test_file.kml")
file <- file.path(getwd(),"test_file.kml")
import_kml2OnX(file = file, token = Sys.getenv("ONX_TOKEN", unset = ""))
## End Not Run

}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
