% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authHelpers.R
\name{onx_api_base}
\alias{onx_api_base}
\title{Helper functions for authenticating and accessing OnX Web Maps}
\usage{
onx_api_base()
}
\description{
Helper functions for authenticating and accessing OnX Web Maps
}
\note{
Some authentication functions were drafted from vibe code sessions with Copilot using an HAR file, but all have been tested, authenticated, and modified to suite package architecture and functionality.
}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
