% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentHelpers.R
\name{onx_list_markups}
\alias{onx_list_markups}
\title{List user markups by type.}
\usage{
onx_list_markups(
  markup_type = c("waypoints", "tracks", "lines", "shapes"),
  limit = 1000,
  token = Sys.getenv("ONX_TOKEN", unset = ""),
  parse = T,
  include.notes = F
)
}
\arguments{
\item{markup_type}{Character one of "waypoints", "tracks", "lines", "shapes".}

\item{limit}{Integer page size.}

\item{token}{Character bearer token.}

\item{parse}{Boolean whether to return 'cleaned' output as a data frame.}

\item{include.notes}{Boolean whether to attempt to include 'notes' column. Warning, if none of the features have notes, then function will fail.}
}
\value{
Parsed JSON response from the OnX markup endpoint.
}
\description{
List user markups by type.
}
