% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authHelpers.R
\name{onx_login_from_token}
\alias{onx_login_from_token}
\title{Create a session object that stores the bearer token and headers.}
\usage{
onx_login_from_token(token, application_id = "hunt")
}
\arguments{
\item{token}{Character bearer token. Derived from \code{getToken}, typically stored in \code{Sys.getenv("ONX_TOKEN")}.}

\item{application_id}{Character application id. Defaults to "hunt"}
}
\value{
List with token, headers, and api base URL.
}
\description{
Create a session object that stores the bearer token and headers.
}
