% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentHelpers.R
\name{onx_profile}
\alias{onx_profile}
\title{Fetch the current profile from the OnX production API.}
\usage{
onx_profile(token = Sys.getenv("ONX_TOKEN", unset = ""), as_df = T)
}
\description{
Fetch the current profile from the OnX production API.
}
