% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentHelpers.R
\name{onx_subscriptions}
\alias{onx_subscriptions}
\title{Helper functions for viewing OnX Web Maps content including files and account information}
\usage{
onx_subscriptions(token = Sys.getenv("ONX_TOKEN", unset = ""), as_df = T)
}
\description{
Helper functions for viewing OnX Web Maps content including files and account information
}
\note{
Some functions were drafted from vibe code sessions with Copilot using an HAR file, but all have been tested, authenticated, and modified to suite package architecture and functionality.

#' Fetch the current subscriptions for the signed-in account.
}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
