% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmz_to_kml_idfg.R
\name{parse_html}
\alias{parse_html}
\title{Functions for converting KMZ files to KML}
\usage{
parse_html(desc)
}
\description{
Unzip KMZ files and resave as KML files. Close but TOO TAILORED, NEED TO GENERALIZE AND ADD USER FUNCTIONALITY FOR DESCRIPTION!!!!
}
\note{
Will work for IDFG Silviculture files (PVZ, YSH, and NavPointID2); but currently too specific for general use.
}
\examples{
## Not Run:
### Convert KMZ Point Files to KML
files <- dir(fp,pattern = ".kmz",full.names = T)
pt_files <- files[stringr::str_detect(files,"PtsForSampling")]
lapply(pt_files,pts_kmz_to_kml)

## Convert KMZ Polygon Files to KML
stand_files <- files[stringr::str_detect(files,"StandsForSampling")]
lapply(stand_files,stands_kmz_to_kml)

# Clean-up
files <- dir(fp,pattern = ".xsl|.png|doc.kml",full.names = T)
file.remove(files)
## End Not Run

}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
