% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markups.R
\name{plot_markups}
\alias{plot_markups}
\title{Plot markup coordinates.}
\usage{
plot_markups(
  markup_type = c("waypoints", "tracks", "lines", "shapes"),
  limit = 1000,
  token = Sys.getenv("ONX_TOKEN", unset = ""),
  interactive = T
)
}
\arguments{
\item{markup_type}{Character one of "waypoints", "tracks", "lines", "shapes".}

\item{limit}{Integer page size.}

\item{token}{Character bearer token.}

\item{interactive}{Boolean whether to return an interactive leaflet plot with 'terra::plet' or a static one with 'terra::plot'.}
}
\value{
A plot visualizing coordinates of markups.
}
\description{
Plot markup coordinates.
}
