% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCredentials.R
\name{setCredentials}
\alias{setCredentials}
\title{Set Credentials for OnX Web Maps}
\usage{
setCredentials(onx_email, onx_password)
}
\arguments{
\item{onx_email}{A string containing a valid OnX account email address.}

\item{onx_password}{A string containing a valid password to an active OnX account.}
}
\description{
Store login credentials for OnX Web Maps to system environment, optimal for privacy.
}
\author{
Robert Ritson \href{mailto:r2j2ritson@gmail.com}{r2j2ritson@gmail.com}
}
